requirements.txt
pandas==1.4.2
Pillow==9.5.0
requests==2.29.0
streamlit==1.14.0
